Fix nano_malloc and nano_free for IA16.
The size/offset field in the padding area is of type long, but it was
set as an int. This fixes the offset setters to use the same syntax
as the getters. This also means that CHUNK_ALIGN needs to be at least
sizeof(long) since otherwise the padding area could be only 2 bytes
and therefore too small to contain a long.